Refactor test project structure and update unit tests.
The 'GameOfLife.Tests' solution was removed and replaced with 'GameOfLife.Test' project. We had some bug letting nuget find the unit tests in the explorer. This restructuring should fix this issue so that we can test correctly.  This included relocating and updating test files for more cohesive project organization. The JsonStorageTests in particular were significantly expanded and improved for better coverage. Additionally, the solution file was updated to reflect these changes.