Refactor and simplify JsonStorage tests
The JsonStorage unit tests have been revised and streamlined. The revision includes removal of a redundant stub class that was previously used for interfacing. The grid serialization and deserialization tests are now more focused and exhibit improved code clarity. Changes to the Grid and JsonStorage classes was necessary to expose certain properties for testing.